Revert "Land #128, crypttlv support"
This reverts commit 11ad211f96ba707981a9ab77dbb850f10ae860ff, reversing
changes made to 4470b6ab2364235fee55f061885d045e545f7643.
