+++ mettle/src/tlv.h
@@ -16,34 +16,16 @@
 #include "buffer_queue.h"
 #include "tlv_types.h"
 
-#define SESSION_GUID_LEN 16
-#define TLV_PREPEND_LEN 24
-#define TLV_MIN_LEN 8
-
 /*
  * TLV Packets
  */
-
-struct tlv_header {
-	int32_t len;
-	uint32_t type;
-} __attribute__((packed));
-
-struct tlv_xor_header {
-	char xor_key[4];
-	uint8_t session_guid[SESSION_GUID_LEN];
-	uint32_t encryption_flags;
-	struct tlv_header tlv;
-} __attribute__((packed));
-
 struct tlv_packet;
-struct tlv_dispatcher;
 
 struct tlv_packet *tlv_packet_new(uint32_t type, int initial_len);
 
 bool tlv_found_first_packet(struct buffer_queue *q);
 
-struct tlv_packet * tlv_packet_read_buffer_queue(struct tlv_dispatcher *td , struct buffer_queue *q);
+struct tlv_packet * tlv_packet_read_buffer_queue(struct buffer_queue *q);
 
 void *tlv_packet_data(struct tlv_packet *p);
 
@@ -108,13 +90,7 @@ void tlv_packet_free(struct tlv_packet *p);
  * TLV Handler
  */
 struct channel;
-struct tlv_encryption_ctx {
-	unsigned char *key;
-	unsigned char *iv;
-
-	uint32_t flag;
-	bool initialized;
-};
+struct tlv_dispatcher;
 
 struct tlv_handler_ctx {
 	const char *method;
@@ -150,8 +126,6 @@ int tlv_dispatcher_process_request(struct tlv_dispatcher *td, struct tlv_packet 
 int tlv_dispatcher_add_handler(struct tlv_dispatcher *td,
 		const char *method, tlv_handler_cb cb, void *arg);
 
-void tlv_dispatcher_add_encryption(struct tlv_dispatcher *td, struct tlv_encryption_ctx *ctx);
-
 int tlv_dispatcher_enqueue_response(struct tlv_dispatcher *td, struct tlv_packet *p);
 
 void * tlv_dispatcher_dequeue_response(struct tlv_dispatcher *td,
@@ -165,6 +139,7 @@ const char *tlv_dispatcher_get_uuid(struct tlv_dispatcher *td, size_t *len);
 
 int tlv_dispatcher_set_uuid(struct tlv_dispatcher *td, char *uuid, size_t len);
 
+#define SESSION_GUID_LEN 16
 const char *tlv_dispatcher_get_session_guid(struct tlv_dispatcher *td);
 int tlv_dispatcher_set_session_guid(struct tlv_dispatcher *td, char *uuid);
